feat: add IdentifySpeaker component and route, remove IdentifySpeakerTab, and update Layout and Body styles
